requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# Database
# psycopg2-binary==2.9.9
supabase==2.3.0

# Machine Learning
# catboost==1.2.2
# scikit-learn==1.3.2
# pandas==2.1.4
# numpy==1.26.2
joblib==1.3.2

# API & Data Processing
python-dotenv==1.0.0
requests==2.31.0

# Utilities
pytz==2023.3

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
